requirements.txt
# Core Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.1.0
celery>=5.4.0
redis>=5.0.0
python-multipart>=0.0.6

# Auth — Clerk JWT verification
PyJWT[crypto]>=2.8.0
python-dotenv>=1.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
python-dateutil>=2.8.0
requests>=2.25.0

# Analysis & NLP
textblob>=0.17.1
scipy>=1.9.0

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0
cloudscraper>=1.2.71
